Fix Warnings

+++ ObjectiveCloudant/Operations/CDTDeleteDocumentOperation.m
@@ -90,7 +90,7 @@
                         [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding];
                     NSString *msg =
                         [NSString stringWithFormat:@"Document deletion failed with %ld %@.",
-                                                   statusCode, json];
+                                                   (long)statusCode, json];
                     NSDictionary *userInfo =
                         @{NSLocalizedDescriptionKey : NSLocalizedString(msg, nil)};
                     error = [NSError errorWithDomain:CDTObjectiveCloudantErrorDomain
