+++ ObjectiveCloudant/Operations/CDTPutDocumentOperation.m
@@ -99,7 +99,7 @@
                         [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding];
                     NSString *msg =
                         [NSString stringWithFormat:@"Document create or update failed with %ld %@.",
-                                                   statusCode, json];
+                                                   (long)statusCode, json];
                     NSDictionary *userInfo =
                         @{NSLocalizedDescriptionKey : NSLocalizedString(msg, nil)};
                     NSError *error =
